requirement.txt
altgraph==0.16.1
appdirs==1.4.4
Appium-Python-Client==1.0.1
appnope==0.1.0
astroid==2.4.1
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.8.1
blinker==1.4
Brotli==1.0.7
bs4==0.0.1
builtwith==1.3.3
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.0
Django==2.2.6
ExifRead==2.1.2
fake-useragent==0.1.11
Faker==2.0.4
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
fonttools==4.12.1
future==0.18.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
html5lib==1.0.1
httpie==1.0.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itchat==1.3.10
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
kaitaistruct==0.8
lazy-object-proxy==1.4.3
ldap3==2.7
lxml==4.3.4
macholib==1.11
MarkupSafe==1.1.1
mccabe==0.6.1
mitmproxy==5.1.1
modulegraph==0.17
mysql-connector-python==8.0.18
mysqlclient==1.4.5
numpy==1.17.3
pandas==1.0.5
parse==1.15.0
parsel==1.5.2
parso==0.5.1
passlib==1.7.2
pdfkit==0.6.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
Protego==0.1.15
protobuf==3.10.0
ptyprocess==0.6.0
publicsuffix2==2.20191221
py2app==0.19
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyee==7.0.2
Pygments==2.4.2
PyHamcrest==1.9.0
PyInstaller==3.5
pylint==2.5.2
pymongo==3.9.0
PyMySQL==0.9.3
pyobjc-core==6.2
pyobjc-framework-Cocoa==6.2
pyobjc-framework-Quartz==6.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pypng==0.0.20
pyppeteer==0.2.2
PyQRCode==1.2.1
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyquery==1.4.1
pytesseract==0.3.4
python-dateutil==2.8.1
python-whois==0.7.2
pytz==2019.3
PyUserInput==0.1.12
queuelib==1.5.0
redis==3.3.11
requests==2.22.0
requests-file==1.5.1
requests-html==0.10.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Scrapy==1.8.0
scrapyd==1.2.1
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SQLAlchemy==1.3.11
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.1
tornado==6.0.3
tqdm==4.46.1
traitlets==4.3.3
Twisted==19.7.0
typed-ast==1.4.1
urllib3==1.25.6
urwid==2.1.0
virtualenv==16.7.7
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
wrapt==1.12.1
wsproto==0.15.0
wxPython==4.0.7
xlwt==1.3.0
zope.interface==4.6.0
zstandard==0.13.0
